requirements.txt
# portal-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyotp==2.9.0
bcrypt==4.1.2
PyJWT==2.8.0
plotly==5.18.0
gunicorn==21.2.0
WeasyPrint>=66.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0